Update scheduler tests
